vehicle-signal-server-spec/0170-set-error.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Vehicle Signal Server Spec Test: get method tests</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-helper.js"></script>

<h3> Set test: Set with invalid path. Error case </h3>
<h3> Set Signal.Drivetrain.Transmission.xxxxxx </h3>

<h3>
<div id="log"></div>
</h3>

<script>

var t = async_test("Set test: Set with invalid path. Error case.");

var vehicle  = new WebSocket(VISS_URL, VISS_SUBPROTO);

vehicle.onopen = function() {
  // request a signal
  var reqId_1 = "reqid-0170";
  var path = "Signal.Drivetrain.Transmission.xxxx";
  var val = 5; //Gear value: -1 to 15

  // process messages from the server
  vehicle.onmessage = t.step_func(
    function(event) {
      var msg = JSON.parse(event.data);
      var res;

      // if requestId matches and msg.value has value, judge as OK.
      if (isSetErrorResponse(reqId_1, msg)) {
        // try get for verification
        res = true;
        addLogSuccess("setErrorResponse received. err=" + msg.error.number);
      } else {
        res = false;
        addLogFailure("setErrorResponse not received.");
      }
      t.step_timeout(function() {
        assert_true(res, "OK: "+path + ": value = " + msg.value);
        vehicle.close();
        t.done();
      }, TIME_FINISH_WAIT);
    }
  );
  // in case server doesn't return response
  t.step_timeout(function() {
    addLogFailure("Fishish by timeout");
    t.step_timeout(function() {
      assert_true(false);
      vehicle.close();
      t.done();
    }, TIME_FINISH_WAIT);
  }, TIME_OUT_TIME);

  addLogMessage("Send set request: path="+path+", value="+val);
  vehicle.send('{"action":"set","path":"'+ path+'"'
              + ' ,"value":"'+val+'"'
              + ' ,"requestId":"'+reqId_1+'"'
              + '}');
}
</script>
